requirements.txt
trafilatura>=1.8.0
beautifulsoup4>=4.12.3
markdownify>=0.12.1
unstructured>=0.14.10
pdfminer.six>=20240706
python-frontmatter>=1.0.1
PyYAML>=6.0.1
requests>=2.32.3
langchain>=0.0.208
langchain-community>=0.2
langchain-openai>=0.1.6
langchain-huggingface>=0.0.6
chromadb>=0.3.28
tiktoken>=0.4.0
openai>=0.27.14
beautifulsoup4>=4.12.3
trafilatura>=1.8.0
markdownify>=0.12.1
unstructured>=0.14.10
pdfminer.six>=20240706
pi_heif>=1.3.3
unstructured_inference>=0.3.6
pdf2image>=1.16.3
pytesseract>=0.3.10
poppler-utils>=0.68.0
pymupdf>=1.22.0
sentence-transformers>=2.5.1
torch>=2.2.0
huggingface-hub>=0.18.1

# Vector stores
chromadb>=0.5
faiss-cpu>=1.8.0.post1  # optionnel si tu veux FAISS